src/app/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
databricks-sdk>=0.40.0
pydantic>=2.0.0
python-multipart>=0.0.9
